rename AutoSubmitFilterForm to FilterLiveForm
